This commit introduces initial support for training and inference on Apple Silicon GPUs using PyTorch's `mps` backend.
- Unified device abstraction in `unsloth/device_type.py` with `is_mps()`,
-  Implementation of PyTorch-native functional fallbacks for core kernels
  (RMSNorm, RoPE, SwiGLU, CrossEntropy, LayerNorm) that trigger on MPS.
- Added macOS-specific quantization path using `torchao` for 4-bit and
  8-bit.
- Temporary mocking of some dependencies used by unsloth-zoo

+++ unsloth/models/qwen3_moe.py
@@ -66,7 +66,7 @@ def Qwen3MoeSparseMoeBlock_fast_forward(self, X, temp_gate = None, temp_up = Non
     routing_weights /= routing_weights.sum(dim = -1, keepdim = True)
     # we cast back to the input dtype
     routing_weights = routing_weights.to(X.dtype)
-    final_X = torch.zeros((bsz * seq_len, hd), dtype = torch.float32, device = X.device)
+    final_X = torch.zeros((bsz * seq_len, hd), dtype = X.dtype, device = X.device)
 
     # One hot encode the selected experts to create an expert mask
     # this will be used to easily index which expert is going to be sollicitated
